+++ _examples/configuration/from-yaml-file/main.go
@@ -14,7 +14,9 @@ func main() {
 	// Good when you have two configurations, one for development and a different one for production use.
 	// If iris.YAML's input string argument is "~" then it loads the configuration from the home directory
 	// and can be shared between many iris instances.
-	app.Listen(":8080", iris.WithConfiguration(iris.YAML("./configs/iris.yml")))
+	cfg := iris.YAML("./configs/iris.yml")
+	addr := cfg.Other["Addr"].(string)
+	app.Listen(addr, iris.WithConfiguration(cfg))
 
 	// or before run:
 	// app.Configure(iris.WithConfiguration(iris.YAML("./configs/iris.yml")))